code translation utils => locale "pt" should attempt to load "pt-BR" (not "pt-PT")
