+++ packages/docusaurus-utils/src/__tests__/__fixtures__/defaultCodeTranslations/fr-FR.json
