+++ packages/docusaurus-utils/src/__tests__/index.test.ts
@@ -6,7 +6,6 @@
  */
 
 import path from 'path';
-import fs from 'fs-extra';
 import {
   fileToPath,
   simpleHash,
@@ -34,7 +33,6 @@ import {
   findFolderContainingFile,
   getFolderContainingFile,
   updateTranslationFileMessages,
-  readDefaultCodeTranslationMessages,
   parseMarkdownHeadingId,
 } from '../index';
 import {sum} from 'lodash';
@@ -719,92 +717,6 @@ describe('updateTranslationFileMessages', () => {
         t3: {message: 'prefix t3 message suffix', description: 't3 desc'},
       },
     });
-  });
-});
-
-describe('readDefaultCodeTranslationMessages', () => {
-  const dirPath = path.resolve(
-    __dirname,
-    '__fixtures__',
-    'defaultCodeTranslations',
-  );
-
-  async function readAsJSON(filename: string) {
-    return JSON.parse(
-      await fs.readFile(path.resolve(dirPath, filename), 'utf8'),
-    );
-  }
-
-  test('for empty locale', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: '',
-        dirPath,
-      }),
-    ).resolves.toEqual({});
-  });
-
-  test('for unexisting locale', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: 'es',
-        dirPath,
-      }),
-    ).resolves.toEqual({});
-  });
-
-  test('for fr but bad folder', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: '',
-        dirPath: __dirname,
-      }),
-    ).resolves.toEqual({});
-  });
-
-  test('for fr', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: 'fr',
-        dirPath,
-      }),
-    ).resolves.toEqual(await readAsJSON('fr.json'));
-  });
-
-  test('for fr_FR', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: 'fr_FR',
-        dirPath,
-      }),
-    ).resolves.toEqual(await readAsJSON('fr_FR.json'));
-  });
-
-  test('for en', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: 'en',
-        dirPath,
-      }),
-    ).resolves.toEqual(await readAsJSON('en.json'));
-  });
-
-  test('for en_US', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: 'en_US',
-        dirPath,
-      }),
-    ).resolves.toEqual(await readAsJSON('en.json'));
-  });
-
-  test('for en_WHATEVER', async () => {
-    await expect(
-      readDefaultCodeTranslationMessages({
-        locale: 'en_WHATEVER',
-        dirPath,
-      }),
-    ).resolves.toEqual(await readAsJSON('en.json'));
   });
 });
